refactor(rust): extract db ddl operations (slice 9/9)
Move collection, view, and database DDL operations into db/ddl.rs:

- create_collection_impl, create_view_impl
- drop_collection_impl, rename_collection_impl
- drop_database_impl, rename_database_impl
- DatabaseRenameResult

Keep crate-root re-exports so command wrappers and tests continue to use the same symbols. Behavior unchanged.

cargo check: passed
cargo test: 60 passed
